+++ packages/chat/src/components/MessageInput/MessageInput.story.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { useState } from 'react';
 
 import { WithDefaultPalette } from '@/storybook/decorators';
 
@@ -22,35 +21,14 @@ const meta: Meta<typeof MessageInput> = {
 
 export default meta;
 
-const MockComponent = ({ initialMessage }: { initialMessage: string }) => {
-  const [message, setMessage] = useState(initialMessage);
-
-  const onSubmit = () => {
-    setMessage('');
-  };
-
+const MockComponent = () => {
   return (
     <div style={{ width: '400px' }}>
-      <MessageInput
-        placeholder="Message..."
-        message={message}
-        onSubmit={onSubmit}
-        onValueChange={(e) => setMessage(e)}
-      />
+      <MessageInput placeholder="Message..." onSubmit={() => Promise.resolve()} />
     </div>
   );
 };
 
 export const Base: Story = {
-  render: () => <MockComponent initialMessage="" />,
-};
-
-export const LongValue: Story = {
-  render: () => (
-    <MockComponent
-      initialMessage={
-        "Howdy folk how ya'll doing out there. This input was lovingly crafted by your favourite design engineering team here at Voiceflow. We hope you enjoy using it as much as we enjoyed making it. If you have any feedback, please let us know. We're always looking to improve our products. Thanks for using Voiceflow!"
-      }
-    />
-  ),
+  render: () => <MockComponent />,
 };