+++ packages/chat/src/components/SystemResponse/state/end.tsx
@@ -1,14 +1,13 @@
-import { /* useContext, */ useEffect } from 'react';
+import { useContext, useEffect } from 'react';
 
-// import { RuntimeStateAPIContext } from '@/contexts';
-// import { SessionStatus } from '@/types';
+import { RuntimeStateAPIContext } from '@/contexts';
+import { SessionStatus } from '@/types';
 
 const EndState: React.FC = () => {
-  // const runtime = useContext(RuntimeStateAPIContext);
+  const runtime = useContext(RuntimeStateAPIContext);
 
   useEffect(() => {
-    // TODO: uncomment this
-    // runtime.setStatus(SessionStatus.ENDED);
+    runtime.setStatus(SessionStatus.ENDED);
   }, []);
 
   return null;