+++ packages/chat/src/views/ChatWidget/styles.css.ts
@@ -1,8 +1,10 @@
 import { style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
-const CHAT_WIDTH = 380;
+export const CHAT_WIDTH = 380;
 const MAX_CHAT_HEIGHT = 800;
+
+export const LAUNCHER_MARGIN = 16;
 
 export const widgetContainer = recipe({
   base: {
@@ -31,9 +33,6 @@ export const widgetContainer = recipe({
 
 export const chatContainer = style({
   width: CHAT_WIDTH,
-  overflow: 'hidden',
-  borderRadius: '$2',
-  boxShadow: '0 2px 48px rgba(19,33,68,0.16), 0 0 0 1px $shadow4',
   maxHeight: MAX_CHAT_HEIGHT,
 
   selectors: {
@@ -60,19 +59,6 @@ export const launcherContainer = style({
   selectors: {
     [`${widgetContainer.classNames.base} &`]: {
       position: 'absolute',
-    },
-
-    [`.${widgetContainer.classNames.variants.withChat.true} > &`]: {
-      opacity: 0,
-      pointerEvents: 'none',
-      transform: 'translateY(100%)',
-      transition: 'transform 300ms cubic-bezier(0.85, 0, 0.6, 1), opacity 150ms linear',
-    },
-    [`.${widgetContainer.classNames.variants.withChat.false} > &`]: {
-      opacity: 1,
-      pointerEvents: 'auto',
-      transform: 'translateY(0%)',
-      transition: 'transform 300ms cubic-bezier(0, 0.95, 0.1, 1), opacity 150ms linear',
     },
   },
 });